+++ glfw/cocoa_init.m
@@ -26,11 +26,6 @@
 
 #include "internal.h"
 #include <sys/param.h> // For MAXPATHLEN
-#if MAC_OS_X_VERSION_MAX_ALLOWED < 101200
-  #define NSEventMaskKeyUp NSKeyUpMask
-  #define NSEventMaskKeyDown NSKeyDownMask
-  #define NSEventModifierFlagCommand NSCommandKeyMask
-#endif
 
 // Change to our application bundle's resources directory, if present
 //
@@ -295,15 +290,20 @@ static GLFWbool initializeTIS(void)
 @end  // GLFWHelper
 
 @interface GLFWApplication : NSApplication
+- (void)tick_callback;
+- (void)render_frame_received:(id)displayIDAsID;
 @end
 
-extern void dispatchCustomEvent(NSEvent *event);
-
 @implementation GLFWApplication
-- (void)sendEvent:(NSEvent *)event {
-    if (event.type == NSEventTypeApplicationDefined) {
-        dispatchCustomEvent(event);
-    } else [super sendEvent:event];
+- (void)tick_callback
+{
+    _glfwDispatchTickCallback();
+}
+
+- (void)render_frame_received:(id)displayIDAsID
+{
+    CGDirectDisplayID displayID = [(NSNumber*)displayIDAsID unsignedIntValue];
+    _glfwDispatchRenderFrame(displayID);
 }
 @end
 
@@ -462,6 +462,7 @@ static GLFWtickcallback tick_callback = NULL;
 static void* tick_callback_data = NULL;
 static bool tick_callback_requested = false;
 
+
 void _glfwDispatchTickCallback() {
     if (tick_callback) {
         tick_callback_requested = false;
@@ -472,7 +473,7 @@ void _glfwDispatchTickCallback() {
 void _glfwPlatformRequestTickCallback() {
     if (!tick_callback_requested) {
         tick_callback_requested = true;
-        _glfwCocoaPostEmptyEvent(TICK_CALLBACK_EVENT_TYPE, 0, false);
+        [NSApp performSelectorOnMainThread:@selector(tick_callback) withObject:nil waitUntilDone:NO];
     }
 }