+++ kitty/fonts.c
@@ -58,6 +58,7 @@ typedef struct {
 
 
 static hb_buffer_t *harfbuzz_buffer = NULL;
+static hb_feature_t no_calt_feature = {0};
 static char_type shape_buffer[4096] = {0};
 static size_t max_texture_size = 1024, max_array_len = 1024;
 
@@ -693,6 +694,7 @@ typedef struct {
 
 typedef struct {
     unsigned int first_glyph_idx, first_cell_idx, num_glyphs, num_cells;
+    bool has_special_glyph;
 } Group;
 
 typedef struct {
@@ -717,7 +719,7 @@ num_codepoints_in_cell(CPUCell *cell) {
 }
 
 static inline void
-shape(CPUCell *first_cpu_cell, GPUCell *first_gpu_cell, index_type num_cells, hb_font_t *font) {
+shape(CPUCell *first_cpu_cell, GPUCell *first_gpu_cell, index_type num_cells, hb_font_t *font, bool disable_ligature) {
     if (group_state.groups_capacity <= 2 * num_cells) {
         group_state.groups_capacity = MAX(128, 2 * num_cells);  // avoid unnecessary reallocs
         group_state.groups = realloc(group_state.groups, sizeof(Group) * group_state.groups_capacity);
@@ -741,7 +743,13 @@ shape(CPUCell *first_cpu_cell, GPUCell *first_gpu_cell, index_type num_cells, hb
     group_state.last_cpu_cell = first_cpu_cell + (num_cells ? num_cells - 1 : 0);
     group_state.last_gpu_cell = first_gpu_cell + (num_cells ? num_cells - 1 : 0);
     load_hb_buffer(first_cpu_cell, first_gpu_cell, num_cells);
-    hb_shape(font, harfbuzz_buffer, NULL, 0);
+
+    if (disable_ligature) {
+        hb_shape(font, harfbuzz_buffer, &no_calt_feature, 1);
+    } else {
+        hb_shape(font, harfbuzz_buffer, NULL, 0);
+    }
+
     unsigned int info_length, positions_length;
     group_state.info = hb_buffer_get_glyph_infos(harfbuzz_buffer, &info_length);
     group_state.positions = hb_buffer_get_glyph_positions(harfbuzz_buffer, &positions_length);
@@ -806,8 +814,8 @@ check_cell_consumed(CellData *cell_data, CPUCell *last_cpu_cell) {
 
 
 static inline void
-shape_run(CPUCell *first_cpu_cell, GPUCell *first_gpu_cell, index_type num_cells, Font *font) {
-    shape(first_cpu_cell, first_gpu_cell, num_cells, harfbuzz_font_for_face(font->face));
+shape_run(CPUCell *first_cpu_cell, GPUCell *first_gpu_cell, index_type num_cells, Font *font, bool disable_ligature) {
+    shape(first_cpu_cell, first_gpu_cell, num_cells, harfbuzz_font_for_face(font->face), disable_ligature);
 #if 0
         // You can also generate this easily using hb-shape --show-extents --cluster-level=1 --shapers=ot /path/to/font/file text
         hb_buffer_serialize_glyphs(harfbuzz_buffer, 0, group_state.num_glyphs, (char*)canvas, sizeof(pixel) * CELLS_IN_CANVAS * cell_width * cell_height, NULL, harfbuzz_font_for_face(font->face), HB_BUFFER_SERIALIZE_FORMAT_TEXT, HB_BUFFER_SERIALIZE_FLAG_DEFAULT | HB_BUFFER_SERIALIZE_FLAG_GLYPH_EXTENTS);
@@ -869,6 +877,7 @@ shape_run(CPUCell *first_cpu_cell, GPUCell *first_gpu_cell, index_type num_cells
     current_group->num_glyphs = 1; \
     current_group->first_glyph_idx = G(glyph_idx); \
 }
+        if (is_special) current_group->has_special_glyph = true;
         if (is_last_glyph) {
             // soak up all remaining cells
             if (G(cell_idx) < G(num_cells)) {
@@ -930,6 +939,23 @@ merge_groups_for_pua_space_ligature() {
 }
 
 static inline void
+split_run_at_offset(index_type cursor_offset, index_type *left, index_type *right) {
+    *left = 0; *right = 0;
+    for (unsigned idx = 0; idx < G(group_idx) + 1; idx++) {
+        Group *group = G(groups) + idx;
+        if (group->first_cell_idx <= cursor_offset && cursor_offset < group->first_cell_idx + group->num_cells) {
+            GPUCell *first_cell = G(first_gpu_cell) + group->first_cell_idx;
+            if (group->num_cells > 1 && group->has_special_glyph && (first_cell->attrs & WIDTH_MASK) == 1) {
+                // likely a calt ligature
+                *left = group->first_cell_idx; *right = group->first_cell_idx + group->num_cells;
+            }
+            break;
+        }
+    }
+}
+
+
+static inline void
 render_groups(FontGroup *fg, Font *font, bool center_glyph) {
     unsigned idx = 0;
     ExtraGlyphs ed;
@@ -968,7 +994,7 @@ test_shape(PyObject UNUSED *self, PyObject *args) {
         FontGroup *fg = font_groups;
         font = fg->fonts + fg->medium_font_idx;
     }
-    shape_run(line->cpu_cells, line->gpu_cells, num, font);
+    shape_run(line->cpu_cells, line->gpu_cells, num, font, false);
 
     PyObject *ans = PyList_New(0);
     unsigned int idx = 0;
@@ -989,11 +1015,28 @@ test_shape(PyObject UNUSED *self, PyObject *args) {
 #undef G
 
 static inline void
-render_run(FontGroup *fg, CPUCell *first_cpu_cell, GPUCell *first_gpu_cell, index_type num_cells, ssize_t font_idx, bool pua_space_ligature, bool center_glyph) {
+render_run(FontGroup *fg, CPUCell *first_cpu_cell, GPUCell *first_gpu_cell, index_type num_cells, ssize_t font_idx, bool pua_space_ligature, bool center_glyph, int cursor_offset) {
     switch(font_idx) {
         default:
-            shape_run(first_cpu_cell, first_gpu_cell, num_cells, &fg->fonts[font_idx]);
+            shape_run(first_cpu_cell, first_gpu_cell, num_cells, &fg->fonts[font_idx], false);
             if (pua_space_ligature) merge_groups_for_pua_space_ligature();
+            else if (cursor_offset > -1) {
+                index_type left, right;
+                split_run_at_offset(cursor_offset, &left, &right);
+                if (right > left) {
+                    if (left) {
+                        shape_run(first_cpu_cell, first_gpu_cell, left, &fg->fonts[font_idx], false);
+                        render_groups(fg, &fg->fonts[font_idx], center_glyph);
+                    }
+                        shape_run(first_cpu_cell + left, first_gpu_cell + left, right - left, &fg->fonts[font_idx], true);
+                        render_groups(fg, &fg->fonts[font_idx], center_glyph);
+                    if (right < num_cells) {
+                        shape_run(first_cpu_cell + right, first_gpu_cell + right, num_cells - right, &fg->fonts[font_idx], false);
+                        render_groups(fg, &fg->fonts[font_idx], center_glyph);
+                    }
+                    break;
+                }
+            }
             render_groups(fg, &fg->fonts[font_idx], center_glyph);
             break;
         case BLANK_FONT:
@@ -1009,13 +1052,21 @@ render_run(FontGroup *fg, CPUCell *first_cpu_cell, GPUCell *first_gpu_cell, inde
 }
 
 void
-render_line(FONTS_DATA_HANDLE fg_, Line *line) {
-#define RENDER if (run_font_idx != NO_FONT && i > first_cell_in_run) render_run(fg, line->cpu_cells + first_cell_in_run, line->gpu_cells + first_cell_in_run, i - first_cell_in_run, run_font_idx, false, center_glyph);
+render_line(FONTS_DATA_HANDLE fg_, Line *line, index_type lnum, Cursor *cursor) {
+#define RENDER if (run_font_idx != NO_FONT && i > first_cell_in_run) { \
+    int cursor_offset = -1; \
+    if (disable_ligature_in_line && first_cell_in_run <= cursor->x && cursor->x <= i) cursor_offset = cursor->x - first_cell_in_run; \
+    render_run(fg, line->cpu_cells + first_cell_in_run, line->gpu_cells + first_cell_in_run, i - first_cell_in_run, run_font_idx, false, center_glyph, cursor_offset); \
+}
     FontGroup *fg = (FontGroup*)fg_;
     ssize_t run_font_idx = NO_FONT;
     bool center_glyph = false;
+    bool disable_ligature_in_line = false;
     index_type first_cell_in_run, i;
     attrs_type prev_width = 0;
+    if (cursor != NULL && OPT(disable_ligatures_under_cursor)) {
+        if (lnum == cursor->y) disable_ligature_in_line = true;
+    }
     for (i=0, first_cell_in_run=0; i < line->xnum; i++) {
         if (prev_width == 2) { prev_width = 0; continue; }
         CPUCell *cpu_cell = line->cpu_cells + i;
@@ -1042,19 +1093,18 @@ render_line(FONTS_DATA_HANDLE fg_, Line *line) {
                 num_spaces++;
                 // We have a private use char followed by space(s), render it as a multi-cell ligature.
                 GPUCell *space_cell = line->gpu_cells + i + num_spaces;
-                // Ensure the space cell uses the foreground/background colors from the PUA cell.
-                // This is needed because there are stupid applications like
-                // powerline that use PUA+space with different foreground colors
+                // Ensure the space cell uses the foreground color from the PUA cell.
+                // This is needed because there are applications like
+                // Powerline that use PUA+space with different foreground colors
                 // for the space and the PUA. See for example: https://github.com/kovidgoyal/kitty/issues/467
                 space_cell->fg = gpu_cell->fg;
-                space_cell->bg = gpu_cell->bg;
                 space_cell->decoration_fg = gpu_cell->decoration_fg;
             }
             if (num_spaces) {
                 center_glyph = true;
                 RENDER
                 center_glyph = false;
-                render_run(fg, line->cpu_cells + i, line->gpu_cells + i, num_spaces + 1, cell_font_idx, true, center_glyph);
+                render_run(fg, line->cpu_cells + i, line->gpu_cells + i, num_spaces + 1, cell_font_idx, true, center_glyph, -1);
                 run_font_idx = NO_FONT;
                 first_cell_in_run = i + num_spaces + 1;
                 prev_width = line->gpu_cells[i+num_spaces].attrs & WIDTH_MASK;
@@ -1246,7 +1296,7 @@ test_render_line(PyObject UNUSED *self, PyObject *args) {
     PyObject *line;
     if (!PyArg_ParseTuple(args, "O!", &Line_Type, &line)) return NULL;
     if (!num_font_groups) { PyErr_SetString(PyExc_RuntimeError, "must create font group first"); return NULL; }
-    render_line((FONTS_DATA_HANDLE)font_groups, (Line*)line);
+    render_line((FONTS_DATA_HANDLE)font_groups, (Line*)line, 0, NULL);
     Py_RETURN_NONE;
 }
 
@@ -1368,6 +1418,12 @@ init_fonts(PyObject *module) {
     harfbuzz_buffer = hb_buffer_create();
     if (harfbuzz_buffer == NULL || !hb_buffer_allocation_successful(harfbuzz_buffer) || !hb_buffer_pre_allocate(harfbuzz_buffer, 2048)) { PyErr_NoMemory(); return false; }
     hb_buffer_set_cluster_level(harfbuzz_buffer, HB_BUFFER_CLUSTER_LEVEL_MONOTONE_CHARACTERS);
+#define feature_str "-calt"
+    if (!hb_feature_from_string(feature_str, sizeof(feature_str) - 1, &no_calt_feature)) {
+        PyErr_SetString(PyExc_RuntimeError, "Failed to create -calt harfbuzz feature");
+        return false;
+    }
+#undef feature_str
     if (PyModule_AddFunctions(module, module_methods) != 0) return false;
     current_send_sprite_to_gpu = send_sprite_to_gpu;
     return true;