+++ kitty/shaders.c
@@ -243,20 +243,25 @@ cell_update_uniform_block(ssize_t vao_idx, Screen *screen, int uniform_buffer, G
         copy_color_table_to_buffer(screen->color_profile, (GLuint*)rd, cell_program_layouts[CELL_PROGRAM].color_table.offset / sizeof(GLuint), cell_program_layouts[CELL_PROGRAM].color_table.stride / sizeof(GLuint));
     }
     // Cursor position
+    enum { BLOCK_IDX = 0, BEAM_IDX = 6, UNDERLINE_IDX = 7, UNFOCUSED_IDX = 8 };
     if (cursor->is_visible) {
         rd->cursor_x = screen->cursor->x, rd->cursor_y = screen->cursor->y;
         if (cursor->is_focused) {
             switch(cursor->shape) {
                 default:
-                    rd->cursor_fg_sprite_idx = 0; break;
+                    rd->cursor_fg_sprite_idx = BLOCK_IDX; break;
                 case CURSOR_BEAM:
-                    rd->cursor_fg_sprite_idx = 6; break;
+                    rd->cursor_fg_sprite_idx = BEAM_IDX; break;
                 case CURSOR_UNDERLINE:
-                    rd->cursor_fg_sprite_idx = 7; break;
+                    rd->cursor_fg_sprite_idx = UNDERLINE_IDX; break;
             }
-        } else rd->cursor_fg_sprite_idx = 8;
+        } else rd->cursor_fg_sprite_idx = UNFOCUSED_IDX;
     } else rd->cursor_x = screen->columns, rd->cursor_y = screen->lines;
-    rd->cursor_w = rd->cursor_x + MAX(1, screen_current_char_width(screen)) - 1;
+    rd->cursor_w = rd->cursor_x;
+    if (
+            (rd->cursor_fg_sprite_idx == BLOCK_IDX || rd->cursor_fg_sprite_idx == UNDERLINE_IDX) &&
+            screen_current_char_width(screen) > 1
+    ) rd->cursor_w += 1;
 
     rd->xnum = screen->columns; rd->ynum = screen->lines;
 
@@ -286,12 +291,20 @@ cell_prepare_to_render(ssize_t vao_idx, ssize_t gvao_idx, Screen *screen, GLfloa
 
     ensure_sprite_map(fonts_data);
 
-    if (screen->scroll_changed || screen->is_dirty) {
+    bool cursor_pos_changed = screen->cursor->x != screen->last_rendered_cursor_x
+                           || screen->cursor->y != screen->last_rendered_cursor_y;
+
+    if (screen->scroll_changed || screen->is_dirty || (OPT(disable_ligatures_under_cursor) && cursor_pos_changed)) {
         sz = sizeof(GPUCell) * screen->lines * screen->columns;
         address = alloc_and_map_vao_buffer(vao_idx, sz, cell_data_buffer, GL_STREAM_DRAW, GL_WRITE_ONLY);
-        screen_update_cell_data(screen, address, fonts_data);
+        screen_update_cell_data(screen, address, fonts_data, OPT(disable_ligatures_under_cursor) && cursor_pos_changed);
         unmap_vao_buffer(vao_idx, cell_data_buffer); address = NULL;
         changed = true;
+    }
+
+    if (cursor_pos_changed) {
+        screen->last_rendered_cursor_x = screen->cursor->x;
+        screen->last_rendered_cursor_y = screen->cursor->y;
     }
 
     if (screen_is_selection_dirty(screen)) {